Integration tests done

+++ src/main/java/com/endava/twitterdemo/Service/Implementation/UserServiceImp.java
@@ -1,6 +1,6 @@
 package com.endava.twitterdemo.Service.Implementation;
 
-import com.endava.twitterdemo.Model.Tweet;
+import com.endava.twitterdemo.Exceptions.UserDoesNotExistException;
 import com.endava.twitterdemo.Model.User;
 import com.endava.twitterdemo.Repository.UserRepository;
 import com.endava.twitterdemo.Service.UserService;
@@ -8,9 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
-import java.util.Date;
 import java.util.Optional;
-import java.util.Set;
 
 @Service
 public class UserServiceImp implements UserService {
@@ -18,17 +16,18 @@ public class UserServiceImp implements UserService {
     private final UserRepository userRepository;
     @Value("${spring.user.password}")
     private String defaultUserPassword;
+
     @Autowired
-    public UserServiceImp(UserRepository userRepository){
+    public UserServiceImp(UserRepository userRepository) {
         this.userRepository = userRepository;
     }
 
     @Override
     public User createUser(User user) {
 
-            if(user.getPassword().equals("")){
-                user.setPassword(defaultUserPassword);
-            }
+        if (user.getPassword().equals("")) {
+            user.setPassword(defaultUserPassword);
+        }
         return userRepository.save(user);
     }
 
@@ -38,14 +37,18 @@ public class UserServiceImp implements UserService {
     }
 
     @Override
-    public User updatePassword(Long id,String password) {
-       Optional<User> user = userRepository.findById(id);
-       user.get().setPassword(password);
-         return userRepository.save(user.get());
+    public User updatePassword(Long id, String password) {
+        Optional<User> user = userRepository.findById(id);
+        user.get().setPassword(password);
+        return userRepository.save(user.get());
     }
 
     @Override
-    public void deleteUser(Long id) {
+    public void deleteUser(Long id) throws UserDoesNotExistException {
+        Optional<User> isHeHere = userRepository.findById(id);
+        if (!isHeHere.isPresent()) {
+            throw new UserDoesNotExistException("User does not exist");
+        }
         userRepository.deleteById(id);
     }
